+++ .travis.yml
@@ -1,3 +1,4 @@
+sudo: false
 language: python
 dist: xenial
 python:
