+++ requirements.txt
@@ -3,12 +3,14 @@ attrs==19.1.0
 beautifulsoup4==4.7.1
 certifi==2019.3.9
 chardet==3.0.4
+coverage==4.5.3
 idna==2.8
 lxml==4.3.3
 more-itertools==7.0.0
 pluggy==0.11.0
 py==1.8.0
 pytest==4.4.1
+pytest-cov==2.7.1
 requests==2.21.0
 six==1.12.0
 soupsieve==1.9.1
